Add owner of bot and complete emo_count

+++ cmds/emo_count.py
@@ -3,6 +3,7 @@ from discord.ext import commands
 from core.classes import Cog_Ext
 
 import os, re, pymongo
+from operator import itemgetter
 
 client = pymongo.MongoClient(
     f"mongodb+srv://Kerati:{os.getenv('MONGO_PASSWORD')}@kerati.o6ymg.mongodb.net/Kerati?retryWrites=true&w=majority"
@@ -13,15 +14,50 @@ coll = db['emoji_counter']
 
 
 class Emo_count(Cog_Ext):
+    @commands.Cog.listener()
+    async def on_ready(self):
+        self.mongo_emojis = [db_emo['_id'] for db_emo in coll.find()]
+        self.guild_emojis = [
+            g_emo.id for g_emo in (
+                await self.bot.fetch_guild(669934356172636199)).emojis
+        ]
+
+        for db_emo in self.mongo_emojis:
+            if db_emo not in self.guild_emojis:
+                coll.delete_one({'_id': db_emo})
+        for g_emo in self.guild_emojis:
+            if g_emo not in self.mongo_emojis:
+                emo = self.bot.get_emoji(g_emo)
+                coll.insert_one({
+                    '_id': g_emo,
+                    'name': emo.name,
+                    'animated': emo.animated,
+                    'count': 0,
+                })
+
+    @commands.Cog.listener()
+    async def on_message(self, msg):
+        if msg.channel.type not in {
+                discord.ChannelType.text, discord.ChannelType.group
+        } or msg.guild.id != 669934356172636199 or msg.author.bot:
+            return
+        msg_emojis = list(set(re.findall(r'<a?:.*?:(\d*)>', msg.content)))
+        for m_emo in msg_emojis:
+            coll.update_one({'_id': int(m_emo)}, {'$inc': {'count': 1}})
+
     @commands.command(aliases=['ecc'])
     async def emo_counter_clear(self, ctx):
         await ctx.message.delete()
+        if not (await self.bot.is_owner(ctx.author)): return
+
         for db_emo in self.mongo_emojis:
             coll.delete_one({'_id': db_emo})
 
     @commands.command(aliases=['ecr'])
     async def emo_counter_reset(self, ctx):
         await ctx.message.delete()
+        if not (await self.bot.is_owner(ctx.author)): return
+
         self.guild_emojis = [
             g_emo.id for g_emo in (
                 await self.bot.fetch_guild(669934356172636199)).emojis
@@ -40,41 +76,56 @@ class Emo_count(Cog_Ext):
             })
         for g_emo in self.guild_emojis:
             if g_emo not in self.mongo_emojis:
+                emo = self.bot.get_emoji(g_emo)
                 coll.insert_one({
                     '_id': g_emo,
-                    'name': self.bot.get_emoji(g_emo).name,
+                    'name': emo.name,
+                    'animated': emo.animated,
                     'count': 0,
                 })
         self.mongo_emojis = [db_emo['_id'] for db_emo in coll.find()]
 
-    @commands.Cog.listener()
-    async def on_ready(self):
-        self.mongo_emojis = [db_emo['_id'] for db_emo in coll.find()]
-        self.guild_emojis = [
-            g_emo.id for g_emo in (
-                await self.bot.fetch_guild(669934356172636199)).emojis
-        ]
+    @commands.command(aliases=['er'])
+    async def emo_rank(self, ctx, arg=None):
+        await ctx.message.delete(delay=3)
+        if not arg: arg = 10
+        else:
+            emo_id = re.search(r'<a?:.*?:(\d*)>', arg)
+            if not emo_id:
+                arg = int(arg)
+        db_emo_list = [{
+            'id': db_emo['_id'],
+            'name': db_emo['name'],
+            'animated': db_emo['animated'],
+            'count': db_emo['count'],
+        } for db_emo in coll.find()]
 
-        for db_emo in self.mongo_emojis:
-            if db_emo not in self.guild_emojis:
-                coll.delete_one({'_id': db_emo})
-        for g_emo in self.guild_emojis:
-            if g_emo not in self.mongo_emojis:
-                coll.insert_one({
-                    '_id': g_emo,
-                    'name': self.bot.get_emoji(g_emo).name,
-                    'count': 0,
-                })
+        emo_rank = sorted(db_emo_list, key=itemgetter('count'), reverse=True)
 
-    @commands.Cog.listener()
-    async def on_message(self, msg):
-        if msg.channel.type not in {
-                discord.ChannelType.text, discord.ChannelType.group
-        } or msg.guild.id != 669934356172636199:
-            return
-        msg_emojis = list(set(re.findall(r'<a?:.*?:(\d*)>', msg.content)))
-        for m_emo in msg_emojis:
-            coll.update_one({'_id': int(m_emo)}, {'$inc': {'count': 1}})
+        if isinstance(arg, int):
+            if arg < 10:
+                arg = 10
+            elif arg > len(emo_rank):
+                arg = len(emo_rank)
+
+            embed = discord.Embed(title=f'排名 {arg - 9} ~ {arg}',
+                                  description='')
+            for i, emo in enumerate(emo_rank[arg - 10:arg], start=1):
+                embed.description += f"<{'a' if emo['animated'] else ''}:{emo['name']}:{emo['id']}>`{emo['count']:^3d}`次　"
+                if i % 5 == 0:
+                    embed.description += '\n\n'
+            await ctx.send(embed=embed, delete_after=30)
+        elif isinstance(emo_id[1], str):
+            index = next((i for (i, e) in enumerate(emo_rank)
+                          if e['id'] == int(emo_id[1])), None)
+            if index:
+                embed = discord.Embed(description=f'''
+                    使用次數：{emo_rank[index]["count"]}
+
+
+                    排名：{index + 1}''')
+                embed.set_thumbnail(url=self.bot.get_emoji(int(emo_id[1])).url)
+                await ctx.send(embed=embed)
 
 
 def setup(bot):
